fix: remove useLikePost hook, move logic inside post footer
