+++ src/components/postsPage/PostFooter.tsx
@@ -1,14 +1,48 @@
-import { type FC } from "react"
+import { type FC, useState } from "react"
 import { Icon } from "../Icon"
 import { Heart } from "../Icons/Heart"
 import Link from "next/link"
 import { type PostWithUser } from "./types"
+import { api } from "~/utils/api"
+import toast from "react-hot-toast"
+import { ParseZodErrorToString } from "~/utils/helpers"
+import { CONFIG } from "~/config"
 
 export const PostFooter: FC<{
 	isLikedByUser: boolean
 	postWithUser: PostWithUser
-	onLike: (action: "like" | "unlike") => void
-}> = ({ isLikedByUser, postWithUser, onLike }) => {
+}> = ({ isLikedByUser, postWithUser }) => {
+	const [isPostLiked, setPostLiked] = useState<boolean>(isLikedByUser)
+	const [postLikedCount, setPostLikedCount] = useState<number>(postWithUser.post.replaysCount)
+
+	const likePost = api.posts.setPostLiked.useMutation({
+		onSuccess: () => {
+			toast.success("Post Liked!")
+			setPostLiked(true)
+			setPostLikedCount(postLikedCount + 1)
+		},
+		onError: (e) => {
+			const error =
+				ParseZodErrorToString(e.data?.zodError) ??
+				"Failed to like post! Please try again later"
+			toast.error(error, { duration: CONFIG.TOAST_ERROR_DURATION_MS })
+		},
+	})
+
+	const unlikePost = api.posts.setPostUnliked.useMutation({
+		onSuccess: () => {
+			toast.success("Post Unliked!")
+			setPostLiked(false)
+			setPostLikedCount(postLikedCount - 1)
+		},
+		onError: (e) => {
+			const error =
+				ParseZodErrorToString(e.data?.zodError) ??
+				"Failed to unlike post! Please try again later"
+			toast.error(error, { duration: CONFIG.TOAST_ERROR_DURATION_MS })
+		},
+	})
+
 	return (
 		<footer className="mt-3 flex text-gray-500">
 			<Link
@@ -27,17 +61,21 @@ export const PostFooter: FC<{
 				className="group flex"
 				onClick={(e) => {
 					e.stopPropagation()
-					onLike(isLikedByUser ? "unlike" : "like")
+					if (isPostLiked) {
+						unlikePost.mutate(postWithUser.post.id)
+					} else {
+						likePost.mutate(postWithUser.post.id)
+					}
 				}}
 			>
 				<Heart
 					className={`h-9 w-9 rounded-full p-1 ${
-						isLikedByUser ? "group-hover:bg-gray-500" : "group-hover:bg-red-500"
+						isPostLiked ? "group-hover:bg-gray-500" : "group-hover:bg-red-500"
 					}`}
-					fillColor={isLikedByUser ? "red" : ""}
+					fillColor={isPostLiked ? "red" : ""}
 				/>
 				<span className={"self-center pl-1 text-xl group-hover:text-red-500"}>
-					{postWithUser.post.likeCount}
+					{postLikedCount}
 				</span>
 			</div>
 		</footer>
